Refactor(tests)
