+++ __tests__/actions/actions.js
@@ -1,4 +1,3 @@
-
 import actions from '../../src/Actions/actions';
 import actionServer from '../../src/Actions/actionServer';
 import authActions from '../../src/Actions/actionsAuth';
@@ -13,26 +12,24 @@ const newsApiGet = Api.get;
 
 newsApiGet.mockReturnValue(Promise.resolve({ data: 'info from api' }));
 
-
-describe('Receive info Actions', () => {
-  test('should receive details from view', () => {
-    actions.receiveDetails();
+describe('Actions', () => {
+  describe('actions', () => {
+    test('should receive details from view', () => {
+      actions.receiveDetails();
+    });
+    test('should make a call to the api', () => {
+      actionServer.receiveArticle();
+      expect(newsApiGet.mock.calls.length).toBe(1);
+    });
   });
-  test('should make a call to the api', () => {
-    actionServer.receiveArticle();
-    expect(newsApiGet.mock.calls.length).toBe(1);
+  describe('authActions', () => {
+    test('gets a user', () => {
+      expect(authActions.getUser());
+    });
   });
-});
-
-describe('Auth Actions', () => {
-  test('gets a user', () => {
-    expect(authActions.getUser());
-  });
-});
-
-
-describe('Logout Actions', () => {
-  it('gets a user', () => {
-    expect(logoutActions.logout());
+  describe('logoutActions', () => {
+    it('logs a user out', () => {
+      expect(logoutActions.logout());
+    });
   });
 });
