+++ __tests__/components/footer.jsx
@@ -3,7 +3,7 @@ import { mount } from 'enzyme';
 import { expect } from 'chai';
 import Footer from '../../src/components/Footer.jsx';
 
-test('Footer', () => {
+describe('Footer', () => {
   const wrapper = mount(<Footer />);
   it('should have a footer', () => {
     expect(wrapper.find('footer')).to.have.length(1);
