+++ __tests__/components/homepage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Homepage from '../../src/components/HomePage';
+import Homepage from '../../src/components/HomePage.jsx';
 
 describe('Welcome', () => {
   const wrapper = shallow(<Homepage />);
