--- __tests__/components/section1-nav.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import { mount } from 'enzyme';
-
-import section from '../../src/components/Body/MainBody.jsx';
-
-describe('Section', () => {
-  const wrapper = mount(<section />);
-  it('renders correctly', () => {
-    const rendered = renderer.create(
-      <section />,
-    );
-  });
-});
