+++ __tests__/components/selectNewsSource.jsx
@@ -2,12 +2,20 @@ import React from 'react';
 import { mount } from 'enzyme';
 import sinon from 'sinon';
 
-import NewsSource from '../../src/components/Body/SelectNewsSource';
+import NewsSource from '../../src/components/Body/SelectNewsSource.jsx';
+
+jest.unmock('../../__mocks__/localStoragemock.js');
+
+const event = {
+  id: 'abc-news-au'
+};
+localStorage.setItem('key', event.id);
 
 
 jest.unmock('../../__mocks__/superagent.js');
 describe('NewsSource', () => {
   const wrapper = mount(<NewsSource />);
+
   test('should have have divs', () => {
     expect(wrapper.find('div').length).toBe(7);
   });
@@ -18,7 +26,7 @@ describe('NewsSource', () => {
     wrapper.instance().componentDidMount();
     expect(spyDidMount.calledOnce).toBe(true);
   });
-  const handleChange = wrapper.instance().handleChange();
+  const handleChange = wrapper.instance().handleChange(event);
   test('handleChange function should be truthy', () => {
     expect(handleChange).toBeTruthy();
   });
