+++ __tests__/components/welcome.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Welcome from '../../src/components/Welcome.jsx';
+import Welcome from '../../src/components/Homepage/Welcome.jsx';
 
 describe('Welcome', () => {
   localStorage.setItem('user', JSON.stringify({ info: {
