+++ __tests__/stores/authStore.js
@@ -21,5 +21,3 @@ describe(' auth Store has correct methods', () => {
     });
   });
 });
-
-
